+++ components/containers/DynamicForm/Settings.tsx
@@ -19,12 +19,11 @@ const handleDelete = async (formID: number) => {
   // redirect to view templates page on success
   const resp = await axios({
     url: "/api/templates",
-    method: "POST",
+    method: "DELETE",
     headers: {
-      "Content-Type": "multipart/form-data",
+      "Content-Type": "application/json",
     },
     data: {
-      method: "DELETE",
       formID: formID,
     },
     timeout: process.env.NODE_ENV === "production" ? 60000 : 0,