+++ lib/integration/crud.ts
@@ -18,25 +18,26 @@ const lambdaClient = new LambdaClient({
 });
 
 async function _crudTemplatesWithCache(payload: CrudTemplateInput): Promise<CrudTemplateResponse> {
-  if (payload.method === "GET" && payload.formID) {
-    const cachedValue = await formCache.formID.check(payload.formID);
+  const { method, formID } = payload;
+  if (method === "GET" && formID) {
+    const cachedValue = await formCache.formID.check(formID);
     if (cachedValue) {
       return cachedValue;
     }
   }
 
   return await _crudTemplates(payload).then((response) => {
-    switch (payload.method) {
+    switch (method) {
       case "GET":
-        if (payload.formID) {
-          formCache.formID.set(payload.formID, response);
+        if (formID) {
+          formCache.formID.set(formID, response);
         }
         break;
+      case "POST":
+      case "PUT":
       case "DELETE":
-      case "UPDATE":
-      case "INSERT":
-        if (payload.formID) {
-          formCache.formID.invalidate(payload.formID);
+        if (formID) {
+          formCache.formID.invalidate(formID);
         }
         break;
       default:
@@ -51,20 +52,20 @@ async function _crudTemplates(payload: CrudTemplateInput): Promise<CrudTemplateR
   const getConfig = (payload: CrudTemplateInput) => {
     const { method, formID, formConfig } = payload;
 
-    switch (payload.method) {
+    switch (method) {
       case "GET":
         return {
-          method,
+          method: "GET",
           formID,
         };
-      case "INSERT":
+      case "POST":
         return {
-          method,
+          method: "INSERT",
           formConfig,
         };
-      case "UPDATE":
+      case "PUT":
         return {
-          method,
+          method: "UPDATE",
           formConfig,
           formID,
         };
@@ -201,13 +202,13 @@ async function _crudOrganizations(
           method,
           organizationID,
         };
-      case "INSERT":
+      case "POST":
         return {
           method,
           organizationNameEn,
           organizationNameFr,
         };
-      case "UPDATE":
+      case "PUT":
         return {
           method,
           organizationID,
@@ -239,7 +240,7 @@ async function _crudOrganizations(
     return timer().then(() => {
       const { method } = payload;
       switch (method) {
-        case "INSERT":
+        case "POST":
           return {
             data: {
               records: [
@@ -251,7 +252,7 @@ async function _crudOrganizations(
               ],
             },
           };
-        case "UPDATE":
+        case "PUT":
           return {
             data: {},
           };