+++ CHANGELOG.md
@@ -23,6 +23,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Renamed `organisation` to `organization` which has an impact on the API access path
+- Modified the middleware functionality and separation of scopes between middlewares
+- A user now needs to have an enabled admin flag (user table) to access the Admin Pages
 
 ### Fixed